TypeError: new_doc() takes 1 positional argument but 2 were given
In the version of the Frappe framework that your server is running, the frappe.new_doc() function is very strict: it can only be used to create an empty document. You are not allowed to pass a dictionary of values to it at the same time.
The correct, version-safe method is a two-step process:
Create the empty document: doc = frappe.new_doc("Account")
Fill in the values: doc.update({ "field1": "value1", ... })
The code I provided was trying to do both steps in one line, which your server's Frappe version does not support.