Updated api.py Please enter Difference Account or set default Stock Adjustment Account for company
The error Please enter Difference Account or set default Stock Adjustment Account for company is coming from ERPNext's core logic. When your script tries to create a Stock Entry (a "Material Receipt") to add inventory, the system requires a special default account to handle any potential discrepancies in stock valuation.
Your setup_company_and_user function was successfully creating all the other accounts but was failing to create this one and, most importantly, failing to set it as the default on the Company document.
The Final, Complete Solution
We will now add this missing logic back into the setup_company_and_user function. This is the last correction needed.